requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
pillow-heif==0.13.0
markdown==3.5.0
